python_service/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sse-starlette
langchain
langchain-groq
langchain-core
langchain-huggingface
faiss-cpu
pypdf
python-multipart
sentence-transformers
langchain-community